Modificación de clases del paquete model

+++ src/java/service/IUsuarioService.java
@@ -18,7 +18,7 @@ public interface IUsuarioService {
 
     public void EliminarRegistro(Usuario usuario);
 
-    public Usuario ObtenerRegistro(int idusuario);
+    public Usuario ObtenerRegistro(long idusuario);
 
     public List<Usuario> ObtenerRegistros();
 }
